Refactor installer fully into installer module
